requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycryptodome==3.15.0
Werkzeug==2.2.2
zipp==3.8.1
